refactor: 컴포넌트 분리로 불필요한 리렌더링 최소화
